requirements.txt
Django==4.2.8
psycopg2-binary==2.9.9
djangorestframework==3.14.0
python-dotenv==1.0.1
ipdb==0.13.13
django-cors-headers==4.3.1
pandas==2.0.3
openpyxl==3.1.2
requests==2.32.2
django-bootstrap5==24.2
djangorestframework-simplejwt==5.3.1